Add new JavaDoc comments on some classes and methods
